tests: split vaults and secrets tests into multiple test files
